refactor: extract project-user-message.js from project.js (PR-12)
Move sticky notes, terminal, context sources, browser extension,
scheduled tasks delegation, and main user message dispatch handlers
into a dedicated module. project.js reduced from 1,857 to 1,358 lines.
